RPS-6123. Refresh Python SDK for Context API. Tests to handle RESOURCE_LOCKED

+++ test/testFiles.py
@@ -396,13 +396,14 @@ class testFilesApi(object):
 
 
 
-        assert_equal(res.data.wordCount, 2)
-        assert_equal(res.data.stringCount, 2)
-        assert_equal(res.data.translationImportErrors, [])
+        if str(status) == '423':
+            assert_equal('RESOURCE_LOCKED', res.code)
+        else:
+            assert_equal(res.data.wordCount, 2)
+            assert_equal(res.data.stringCount, 2)
+            assert_equal(res.data.translationImportErrors, [])
 
-        res, status = self.files_api.deleteUploadedSourceFile(self.uri_import)
-        assert_equal(200, status)
-        assert_equal(self.CODE_SUCCESS_TOKEN, res.code)
+            res, status = self.files_api.deleteUploadedSourceFile(self.uri_import)
 
 
         assert_equal(True, status in [200,202])
